% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{arith_mean_tfidf}
\alias{arith_mean_tfidf}
\title{Arithmetic mean of the two raw TF-IDF views}
\usage{
arith_mean_tfidf(data_matrix)
}
\arguments{
\item{data_matrix}{Binary sparse drug x feature matrix}
}
\value{
Sparse matrix
}
\description{
S = (W_D + W_G) / 2
}
\keyword{internal}
