% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{binary_from_weighted}
\alias{binary_from_weighted}
\title{Recover binary drug x gene-direction matrix from a weighted sparse matrix}
\usage{
binary_from_weighted(weighted_matrix)
}
\arguments{
\item{weighted_matrix}{Sparse matrix (e.g. existing ctfidf)}
}
\value{
Binary dgCMatrix with the same dimnames
}
\description{
Harmonizome-style corpora are ternary (-1/0/+1) encoded as one-hot
gene-direction columns. Nonzero pattern of a TF-IDF matrix recovers that
binary support.
}
\keyword{internal}
