% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIRCUS.R
\name{CIRCUS}
\alias{CIRCUS}
\title{CIRCUS (CIrcular RNA Coordinate Update System)}
\usage{
CIRCUS(ref_gpf_path, ref_path, bed_path, bed6 = FALSE)
}
\arguments{
\item{ref_gpf_path}{A txt file of two columns with no headers: first column with gene names,
second column with transcript IDs. It should match all transcripts in ref_path.
See 'transcript2gene.txt' in test data.}

\item{ref_path}{BED 12 file with no headers of reference transcripts for annotation. Must
contain all 12 columns indicating exon composition of each transcript. 4th column should
contain transcript IDs. see 'annotation.bed' in test data.}

\item{bed_path}{BED file with circRNAs to convert into circRNA UID. 4th column should
contain an original identifier of choice (e.g., circRNA_1). Indicate whether this file is
BED12 or BED6 with the bed6 parameter.}

\item{bed6}{Boolean (T/F) indicator of whether circRNA BED file is BED12 or BED6. Note:
When BED6 input is used, output circRNA UIDs will only contain information about the 5'
and 3' terminal exon blocks (e.g., circZFAND6(3,5)) without intermediate exons (e.g., a
BED12 input would result in circZFAND6(3,4L,5)).}
}
\value{
Data frame with circRNA UID and annotation information of input circRNAs.
"isoform_index" indicates alphabetized indices in the case of duplicated UIDs of circRNA
isoforms that share same exon/intron anatomy but with different splice sites.
}
\description{
Converts BED12 or BED16 coordinates of circular RNAs into Universal Identifiers (UID)
according to nomenclature proposed by Chen et al. (PMID: 36658223).
}
\examples{
library(CIRCUS)
test_bed12 <- CIRCUS(ref_gpf_path = system.file("extdata",'transcript2gene.txt',
package="CIRCUS"),ref_path = system.file("extdata",'annotation.bed',package="CIRCUS"),
bed_path = system.file("extdata",'bed12_coords.bed',package="CIRCUS"),bed6=FALSE)
test_bed6 <- CIRCUS(ref_gpf_path = system.file("extdata",'transcript2gene.txt',
package="CIRCUS"),ref_path = system.file("extdata",'annotation.bed',package="CIRCUS"),
bed_path = system.file("extdata",'bed6_coords.bed',package="CIRCUS"),bed6=TRUE)
}
