% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{col_zscore_sparse}
\alias{col_zscore_sparse}
\title{Column-contextual z-score among observed (nonzero) weights}
\usage{
col_zscore_sparse(W)
}
\arguments{
\item{W}{Sparse numeric matrix}
}
\value{
Sparse matrix of column-contextual z-scores on original support
}
\description{
For each gene-direction column, mu/sd over nonzero drug weights.
}
\keyword{internal}
