% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{combined_tfidf}
\alias{combined_tfidf}
\title{Combined (bidirectional) TF-IDF via Hadamard product of dual views}
\usage{
combined_tfidf(data_matrix)
}
\arguments{
\item{data_matrix}{Binary (or count) sparse drug x feature matrix}
}
\value{
Sparse combined TF-IDF matrix
}
\description{
For binary input:
ctfidf[d,g] = (1/(s_d * df_g)) * log(D/df_g) * log(T/s_d)
}
\details{
This is element-wise, not a matrix product of tfidf(dtm) and t(tfidf(dtm)).
}
\keyword{internal}
