% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compendium_selection.R
\name{compendium_selection}
\alias{compendium_selection}
\title{Compendium selection}
\usage{
compendium_selection()
}
\value{
An integer.
}
\description{
This function will make the selection of which data source to use to run DRUID against. All the datasets were generated with the companion package `cauldron`.
}
