% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Vector cosine similarity}
\usage{
cosine_similarity(query_vector, tfidf_matrix, tfidf_crossprod_mat)
}
\arguments{
\item{query_vector}{Numeric vector (or 1 x G matrix) of length ncol(tfidf_matrix)}

\item{tfidf_matrix}{Sparse (or dense) drug x feature matrix}

\item{tfidf_crossprod_mat}{Row squared-norms of \code{tfidf_matrix}}
}
\value{
Numeric vector of cosine similarities (one per row)
}
\description{
Cosine similarity between a query vector and each *row* of a TF-IDF matrix.
}
\keyword{internal}
