% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprod_matrix.R
\name{crossprod_matrix}
\alias{crossprod_matrix}
\title{Cross-product matrix for tf-idf}
\usage{
crossprod_matrix(tfidf_matrix)
}
\arguments{
\item{tfidf_matrix}{Sparse tf-idf matrix}
}
\value{
cross-product of tf-idf matrix
}
\description{
Returns the cross-product for an input tf-idf matrix
}
