% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{drug_tfidf}
\alias{drug_tfidf}
\title{Drug-as-document TF-IDF}
\usage{
drug_tfidf(data_matrix)
}
\arguments{
\item{data_matrix}{Binary (or count) sparse drug x feature matrix}
}
\value{
Sparse TF-IDF matrix
}
\description{
For binary input: tfidf[d,g] = (1/s_d) * log(D / df_g)
}
\keyword{internal}
