% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\docType{package}
\name{DRUID-package}
\alias{DRUID}
\alias{DRUID-package}
\title{DRUID: Drug Indication Discoverer}
\description{
Rank compounds whose transcriptional signatures mimic or revert a query gene expression profile. Uses dual TF-IDF corpus weighting (default geometric mean), cosine similarity, and an empirical null model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diogocamacho/druid}
  \item Report bugs at \url{https://github.com/diogocamacho/druid/issues}
}

}
\author{
\strong{Maintainer}: Diogo Camacho \email{diogo.camacho@wyss.harvard.edu}

}
\keyword{internal}
