% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\docType{package}
\name{DRUID}
\alias{DRUID}
\alias{DRUID-package}
\title{DRUID: DRUg Indication Discoverer}
\description{
The DRUID package allows one to identify compounds for mimicking or reverting biological profiles of interest.
}
