% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{gene_tfidf}
\alias{gene_tfidf}
\title{Gene-as-document TF-IDF, aligned to drug x gene layout}
\usage{
gene_tfidf(data_matrix)
}
\arguments{
\item{data_matrix}{Binary sparse drug x feature matrix}
}
\value{
Sparse TF-IDF matrix (same dims as data_matrix)
}
\description{
For binary input: tfidf[d,g] = (1/df_g) * log(T / s_d)
}
\keyword{internal}
