% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{geom_mean_tfidf}
\alias{geom_mean_tfidf}
\title{Geometric mean of the two TF-IDF views}
\usage{
geom_mean_tfidf(data_matrix)
}
\arguments{
\item{data_matrix}{Binary sparse drug x feature matrix}
}
\value{
Sparse matrix
}
\description{
S = sqrt(W_D * W_G); softer than Hadamard product, still multiplicative.
}
\keyword{internal}
