% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{global_zscore_sparse}
\alias{global_zscore_sparse}
\title{Global z-score over all nonzero TF-IDF weights in the matrix}
\usage{
global_zscore_sparse(W)
}
\arguments{
\item{W}{Sparse numeric matrix}
}
\value{
Sparse matrix with globally z-scored nonzero values
}
\description{
Global z-score over all nonzero TF-IDF weights in the matrix
}
\keyword{internal}
