% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_corpus.R
\name{normalize_dataset_selection}
\alias{normalize_dataset_selection}
\title{Normalize DRUID dataset selection to an integer id and name}
\usage{
normalize_dataset_selection(selection)
}
\arguments{
\item{selection}{Integer 1-6, or character name (cmap, lincs, ...), or "all"}
}
\value{
list(id=, name=)
}
\description{
Normalize DRUID dataset selection to an integer id and name
}
\keyword{internal}
