% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_weighting.R
\name{stouffer_tfidf}
\alias{stouffer_tfidf}
\title{Stouffer fuse of dual TF-IDF views with row/column contextual z-scores}
\usage{
stouffer_tfidf(data_matrix)
}
\arguments{
\item{data_matrix}{Binary sparse drug x feature matrix}
}
\value{
Sparse Stouffer-combined score matrix
}
\description{
z_D = row-zscore(drug TF-IDF), z_G = col-zscore(gene TF-IDF),
S = (z_D + z_G) / sqrt(2). NLP association strengths; Stouffer combine.
}
\keyword{internal}
