% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Vector cosine similarity}
\usage{
cosine_similarity(query_vector, tfidf_matrix, tfidf_crossprod_mat)
}
\arguments{
\item{query_vector}{A vector to compare to the tf-idf matrix. Length of vector is equal to number of columns of tf-idf matrix}

\item{tfidf_matrix}{Sparse tf-idf matrix (computed from \code{\link{ctfidf}})}

\item{tfidf_crossprod_mat}{The cross-product matrix for the tf-idf matrix (computed with \code{\link{crossprod_matrix}})}
}
\value{
Cosine similarity between query vector and each *row* in tf-idf matrix
}
\description{
Returns the cosine similarity between 2 vectors
}
