% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfidf.R
\name{ctfidf}
\alias{ctfidf}
\title{Combined tf-idf matrix}
\usage{
ctfidf(data_matrix)
}
\arguments{
\item{data_matrix}{Sparse document-term matrix (condition by gene matrix), named (rows and columns)}
}
\value{
A sparse matrix with the computed combined tf-idf
}
\description{
Returns a tf-idf sparse matrix given a condition-gene matrix.
This matrix is a DxT matrix, where D is the set of conditions (disease indications) T are the genes affecting that condition.
To correct effect/prevalence of genes in a given condition or to correct conditions affected by many genes, we compute the tf-idf matrix twice, to get the combined tf-idf:
}
\details{
\eqn{ctfidf = tfidf(dtm) \times t(tfidf(dtm))}
}
