% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{Druid}
\alias{Druid}
\title{Druid}
\description{
A soft wrapper around \link[log:Logger]{log::Logger} to easily
log messages for ambiorix applications.
}
\section{Super class}{
\code{\link[log:Logger]{log::Logger}} -> \code{Druid}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Druid$new()}}
\item \href{#method-split}{\code{Druid$split()}}
\item \href{#method-log}{\code{Druid$log()}}
\item \href{#method-clone}{\code{Druid$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="log" data-topic="Logger" data-id="date">}\href{../../log/html/Logger.html#method-date}{\code{log::Logger$date()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="log" data-topic="Logger" data-id="dir">}\href{../../log/html/Logger.html#method-dir}{\code{log::Logger$dir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="log" data-topic="Logger" data-id="dump">}\href{../../log/html/Logger.html#method-dump}{\code{log::Logger$dump()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="log" data-topic="Logger" data-id="flag">}\href{../../log/html/Logger.html#method-flag}{\code{log::Logger$flag()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="log" data-topic="Logger" data-id="hook">}\href{../../log/html/Logger.html#method-hook}{\code{log::Logger$hook()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="log" data-topic="Logger" data-id="time">}\href{../../log/html/Logger.html#method-time}{\code{log::Logger$time()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="log" data-topic="Logger" data-id="unix">}\href{../../log/html/Logger.html#method-unix}{\code{log::Logger$unix()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Druid$new(
  prefix = ">",
  http_accept = FALSE,
  http_accept_encoding = FALSE,
  http_accept_language = FALSE,
  http_cache_control = FALSE,
  http_connection = FALSE,
  http_cookie = FALSE,
  http_host = FALSE,
  http_sec_fetch_dest = FALSE,
  http_sec_fetch_mode = FALSE,
  http_sec_fetch_site = FALSE,
  http_sec_fetch_user = FALSE,
  http_upgrade_insecure_requests = FALSE,
  http_user_agent = FALSE,
  path_info = FALSE,
  query_string = FALSE,
  remote_addr = FALSE,
  remote_port = FALSE,
  request_method = FALSE,
  script_name = FALSE,
  server_name = FALSE,
  server_port = FALSE,
  content_length = FALSE,
  content_type = FALSE,
  http_referer = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{Prefix of the log messages.}

\item{\code{http_accept, http_accept_encoding, http_accept_language, http_cache_control, http_connection, http_cookie, http_host, http_sec_fetch_dest, http_sec_fetch_mode, http_sec_fetch_site, http_sec_fetch_user, http_upgrade_insecure_requests, http_user_agent, path_info, query_string, remote_addr, remote_port, request_method, script_name, server_name, server_port, content_length, content_type, http_referer}}{Whether to log these headers.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Constructor
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split"></a>}}
\if{latex}{\out{\hypertarget{method-split}{}}}
\subsection{Method \code{split()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Druid$split()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Split every item on its own log line
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Druid$log(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{A request, see \link[ambiorix:Request]{ambiorix::Request}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Log from request
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Druid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
