% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_probability.R
\name{random_probability}
\alias{random_probability}
\title{Random probability}
\usage{
random_probability(similarity_results, gs_size, num_sets, target_tfidf,
  tfidf_crossprod_mat)
}
\arguments{
\item{similarity_results}{Vector of cosine similarities}

\item{gs_size}{Size of the query set}

\item{num_sets}{Number of random sets to generate}

\item{target_tfidf}{Pathway tf-idf matrix.}

\item{tfidf_crossprod_mat}{Cross-product matrix for pathway tf-idf.}
}
\value{
A vector of random probabilities for each pathway given the gene set size.
}
\description{
This function computes the cosine similarity between the set of randomized gene sets and the pathway tf-idf matrix. It calls the cosine_similarity function and returns the probability of a given cosine similarity being random.
}
