% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_druid.R
\name{run_druid}
\alias{run_druid}
\title{Run DRUID}
\usage{
run_druid(dge_matrix, druid_direction, fold_thr, pvalue_thr, entrez,
  num_random, selection)
}
\arguments{
\item{dge_matrix}{This is a 2 column matrix for gene expression changes, where column 1 is the gene fold change and column 2 is the corresponding p-value for the fold change. NOTE: Use log2 of the fold changes as output, for example, from `limma` or `DESeq2`.}

\item{druid_direction}{Desired effect for DRUID to run on: "pos" mimics query phenotype, "neg" reverts query phenotype. Defaults to "neg".}

\item{fold_thr}{Threshold for the fold change to be considered. Defaults to 0 (i.e., log2(1), where fold change is not used as filter)}

\item{pvalue_thr}{Threshold for the p-value of the fold change to be considered. Defaults to 0.05.}

\item{entrez}{EntrezIDs for genes in differentially expressed set. Must be same order as the input matrix.}

\item{num_random}{Number of random sets to be generated to calculate significance of enrichment.  Defaults to 1,000.}

\item{selection}{EntrezIDs for genes in differentially expressed set. Must be same order as the input matrix.}

\item{tfidf_matrix}{tf-idf matrix drug-gene matrix. Column names are Entrez IDs. Computed with \code{\link{ctfidf}}}
}
\value{
A data frame that is sorted on the DRUID score.
}
\description{
Given a selection on the drug compendium, produces a data frame with the results of DRUID. Makes the `concoct` function somewhat redundant (to be revised later.)
}
